# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for MedApp Application",
        "title": "Med App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/visits": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all visits information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Get All visits",
                "operationId": "get-all-visits",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getAllVisitsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a patient visit into database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Create visit",
                "operationId": "create-visit",
                "parameters": [
                    {
                        "description": "Visit main info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medapp.Visit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.newPersonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    }
                }
            }
        },
        "/api/visits/:id": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get visit information by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Get visit By Id",
                "operationId": "get-visit-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Visit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/medapp.VisitOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update visit information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Update visit",
                "operationId": "update-visit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Visit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New visit info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medapp.UpdVisit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.statusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete visit information from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Delete visit",
                "operationId": "delete-visit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Visit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.statusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.errMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.getAllVisitsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/medapp.VisitOutput"
                    }
                }
            }
        },
        "handler.newPersonResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "handler.statusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "medapp.UpdVisit": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string",
                    "example": "2006-01-25"
                },
                "login": {
                    "type": "string",
                    "example": "DrBykov"
                },
                "name": {
                    "type": "string",
                    "example": "Mihail"
                },
                "surname": {
                    "type": "string",
                    "example": "Kravcov"
                },
                "visitdate": {
                    "type": "string",
                    "example": "2021-12-26"
                }
            }
        },
        "medapp.Visit": {
            "type": "object",
            "required": [
                "birthdate",
                "login",
                "name",
                "surname",
                "visitdate"
            ],
            "properties": {
                "birthdate": {
                    "type": "string",
                    "example": "2000-05-13"
                },
                "login": {
                    "type": "string",
                    "example": "DrHouse"
                },
                "name": {
                    "type": "string",
                    "example": "Igor"
                },
                "surname": {
                    "type": "string",
                    "example": "Vasilev"
                },
                "visitdate": {
                    "type": "string",
                    "example": "2022-12-26"
                }
            }
        },
        "medapp.VisitOutput": {
            "type": "object",
            "required": [
                "date",
                "docname",
                "docsurname",
                "patientname",
                "patientsurname"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2022-12-26"
                },
                "docname": {
                    "type": "string",
                    "example": "Gregory"
                },
                "docsurname": {
                    "type": "string",
                    "example": "House"
                },
                "patientname": {
                    "type": "string",
                    "example": "Igor"
                },
                "patientsurname": {
                    "type": "string",
                    "example": "Vasiliev"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}